improved command output handling
Now command output is not sent round trip through the calling instance back to the host (to the checksum file). It's now
directly written to logfiles that are merged into the checksum file finally.

This required some changes to the SUDO handling. This is now done in the SSH provisioning adapter and no longer in the
commands.