improved command interface: fail early
Added better handling of situations where arbitrary many commands can be given (like for And, or Or). Those commands now
have single command as input plus an arbitrary long list of commands. The consequence is that the problem will be
detected at compile time, not runtime (i.e. the panic was removed).