fixed problem with setting hostname
TL;DR One more command to prevent problems with sudo and the hostname not being resolvable.

In the process of setting the hostname multiple steps are required. Each uses sudo (if required). sudo will verify the
currently set hostname. This resulted in problems when then new hostname was set in /etc/hostname, but not made
available in /etc/hosts. To prevent any problems of either the new or old hostname not being known the default hostname
'localhost' is set initially and later overwritten with the new hostname.