fixed problem with ls returning error
If no files with the matcher "*.done" where available and error popped up (very annoying on the first run). The
mechanism with the for loop will prevent this problem. Only problem is that it is executed in a subshell so the $file
variable must be escaped to be handled in the subshell and not in the top level shell.